refactor: new file for the component
