--- packages/shared/src/components/squads/settings/common.ts
@@ -1,13 +1,0 @@
-import classed from '../../../lib/classed';
-
-export const StatusDescription = classed(
-  'p',
-  'typo-subhead text-text-tertiary',
-);
-
-export enum SquadStatus {
-  InProgress = 'in-progress',
-  Pending = 'pending',
-  Rejected = 'rejected',
-  Approved = 'approved',
-}
